third_party/blink/web_tests/fast/events/touch/multi-touch-timestamp.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/resources/testdriver-actions.js"></script>
<div id="target" style="position:absolute; left:0px; right:0px; width:100%; height:100%; background:green">Target</div>
<script>
    var last_event = null;

    function handleTouchEvents(e) {
      if (last_event) {
        t.step(function () {
          assert_greater_than(e.timeStamp, last_event.timeStamp)
        });
      }
      last_event = e;
    }

    function endTest(e) {
      if (!e.touches.length) {
        t.done();
      }
    }

    let t = async_test("Test touch event timestamp with multiple finger.");

    target.addEventListener("touchstart", handleTouchEvents);
    target.addEventListener("touchmove", handleTouchEvents);
    target.addEventListener("touchend", endTest);

    new test_driver.Actions()
            .addPointer("pointer1", "touch")
            .addPointer("pointer2", "touch")
            .pointerMove(0, 0, {origin: target, sourceName: "pointer1"})
            .pointerDown({sourceName: "pointer1"})
            .pointerMove(10, 10, {origin: target, sourceName: "pointer1"})
            .pointerMove(10, 10, {origin: target, sourceName: "pointer2"})
            .pointerDown({sourceName: "pointer2"})
            .addTick()
            .pointerMove(20, 20, {origin: target, sourceName: "pointer2"})
            .addTick()
            .pointerMove(30, 30, {origin: target, sourceName: "pointer1"})
            .addTick()
            .pointerMove(40, 40, {origin: target, sourceName: "pointer2"})
            .addTick()
            .pointerUp({sourceName: "pointer2"})
            .pointerUp({sourceName: "pointer1"})
            .send();
</script>
